factor com utils from word viewer
